+++ src/component/details/userDetail.js
@@ -2,18 +2,18 @@ import React from 'react';
 import {useSelector} from 'react-redux';
 import { useParams } from 'react-router-dom';
 import {
-    getOneConcierge,
+    getOneUsers,
     //conciergesListDate
-} from '../slice/conciergeSlice'
-export const ConsciergeDetail = () => {
+} from '../slice/userSlice'
+export const UserDetail = () => {
 
     const {id} = useParams()
-    const concierge = useSelector((state) => getOneConcierge(state, id))
+    //const concierge = useSelector((state) => getOneUsers(state, id))
     
 
     //const c = useSelector(conciergesListDate())
 
-    console.log(concierge)
+    //console.log(concierge)
     //console.log(c)
 
     return(
